requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
prophet>=1.1
torch>=2.1
torchvision>=0.16
transformers>=4.35
sentencepiece>=0.1.99
stable-baselines3>=2.2
gymnasium>=0.29
shap>=0.43
plotly>=5.18
altair>=5.2
sqlalchemy>=2.0
python-dotenv>=1.0
alpha-vantage>=2.3
matplotlib>=3.8
seaborn>=0.13
networkx>=3.2
